breaking server logic, separating router & controllers
