+++ src/routers/api/sessions.api.js
@@ -1,6 +1,7 @@
 import CustomRouter from "../CustomRouter.js";
 import passport from "../../middlewares/passport.mid.js";
 import passportCb from "../../middlewares/passportCb.mid.js";
+import { register, login, logout, online, googleCallback } from "../../controllers/sessions.controller.js";
 
 class SessionsRouter extends CustomRouter {
   init() {
@@ -20,55 +21,5 @@ class SessionsRouter extends CustomRouter {
   }
 }
 
-async function register(req, res, next) {
-  try {
-    return res.suc200mes("User successfully registered");
-  } catch (err) {
-    return next(err);
-  }
-}
-async function login(req, res, next) {
-  try {
-    return res
-      .cookie("token", req.user.token, { signedCookie: true })
-      .suc200mes("Session Initialized");
-  } catch (err) {
-    return next(err);
-  }
-}
-async function online(req, res, next) {
-  try {
-    if (req.user && req.user.online) {
-      // req.user brought from the payload extracted from the token in the 'jwt' midd.
-      return res.suc200online(req.user._id, req.user.email, req.user.role);
-    } else {
-      return res.err401mes("Offline");
-    }
-  } catch (err) {
-    return next(err);
-  }
-}
-async function logout(req, res, next) {
-  try {
-    if (req.user.online) {
-      return res
-        .clearCookie("token")
-        .suc200mes("Session Closed");
-    } else {
-      return res.err401mes("No session opened");
-    }
-  } catch (err) {
-    return next(err);
-  }
-}
-async function googleCallback(req, res, next) {
-  try {
-    return res.suc200mes("Logged In With Google")
-  } catch (err) {
-    return next(err);
-  }
-}
-
 const sessionsRouter = new SessionsRouter();
-
 export default sessionsRouter.getRouter();
